Put figures in separate folder
