+++ webviz_subsurface/plugins/_rft_plotter/figures/_misfit_figure.py
